+++ docs/examples/security/using_abstract_authentication_middleware.py
@@ -2,15 +2,18 @@ from dataclasses import dataclass
 from typing import Any
 
 import anyio
-from docs.examples.security.using_abstract_authentication_middleware.authentication_middleware import (
-    CustomAuthenticationMiddleware,
-)
 
 from litestar import Litestar, MediaType, Request, Response, WebSocket, get, websocket
+from litestar.connection import ASGIConnection
 from litestar.datastructures import State
 from litestar.di import Provide
-from litestar.exceptions import NotFoundException
+from litestar.exceptions import NotAuthorizedException, NotFoundException
+from litestar.middleware import AbstractAuthenticationMiddleware, AuthenticationResult
 from litestar.middleware.base import DefineMiddleware
+
+API_KEY_HEADER = "X-API-KEY"
+
+TOKEN_USER_DATABASE = {"1": "user_authorized"}
 
 
 @dataclass
@@ -21,6 +24,23 @@ class MyUser:
 @dataclass
 class MyToken:
     api_key: str
+
+
+class CustomAuthenticationMiddleware(AbstractAuthenticationMiddleware):
+    async def authenticate_request(self, connection: ASGIConnection) -> AuthenticationResult:
+        """Given a request, parse the request api key stored in the header and retrieve the user correlating to the token from the DB"""
+
+        # retrieve the auth header
+        auth_header = connection.headers.get(API_KEY_HEADER)
+        if not auth_header:
+            raise NotAuthorizedException()
+
+        # this would be a database call
+        token = MyToken(api_key=auth_header)
+        user = MyUser(name=TOKEN_USER_DATABASE.get(token.api_key))
+        if not user.name:
+            raise NotAuthorizedException()
+        return AuthenticationResult(user=user, auth=token)
 
 
 @get("/")
