+++ docs/usage/security/abstract-authentication-middleware.rst
@@ -47,15 +47,16 @@ Since the above is quite hard to grasp in the abstract, let us see an example.
 We start off by creating a user model. It can be implemented using msgspec, Pydantic, an ODM, ORM, etc.
 For the sake of the example here let us say it is a dataclass:
 
-.. literalinclude:: /examples/security/using_abstract_authentication_middleware/main.py
-    :lines: 16-23
+.. literalinclude:: /examples/security/using_abstract_authentication_middleware.py
+    :lines: 19-26
     :language: python
     :caption: user and token models
 
 
 We can now create our authentication middleware:
 
-.. literalinclude:: /examples/security/using_abstract_authentication_middleware/authentication_middleware.py
+.. literalinclude:: /examples/security/using_abstract_authentication_middleware.py
+    :lines:  29-43
     :language: python
     :caption: authentication_middleware.py
 
@@ -63,8 +64,8 @@ We can now create our authentication middleware:
 Finally, we need to pass our middleware to the Litestar constructor:
 
 
-.. literalinclude:: /examples/security/using_abstract_authentication_middleware/main.py
-    :lines: 60-69
+.. literalinclude:: /examples/security/using_abstract_authentication_middleware.py
+    :lines: 80-88
     :language: python
     :caption: main.py
 
@@ -72,15 +73,15 @@ Finally, we need to pass our middleware to the Litestar constructor:
 That is it. ``CustomAuthenticationMiddleware`` will now run for every request, and we would be able to access these in a
 http route handler in the following way:
 
-.. literalinclude:: /examples/security/using_abstract_authentication_middleware/main.py
-    :lines: 26-31
+.. literalinclude:: /examples/security/using_abstract_authentication_middleware.py
+    :lines: 46-51
     :language: python
     :caption: Accessing the user and auth in a http route handler with ``CustomAuthenticationMiddleware``
 
 Or for a websocket route:
 
-.. literalinclude:: /examples/security/using_abstract_authentication_middleware/main.py
-    :lines: 34-39
+.. literalinclude:: /examples/security/using_abstract_authentication_middleware.py
+    :lines: 54-59
     :language: python
     :caption: Accessing the user and auth in a websocket route handler with ``CustomAuthenticationMiddleware``
 
@@ -88,14 +89,14 @@ Or for a websocket route:
 And if you would like to exclude individual routes outside those configured:
 
 
-.. literalinclude:: /examples/security/using_abstract_authentication_middleware/main.py
-    :lines: 42-50
+.. literalinclude:: /examples/security/using_abstract_authentication_middleware.py
+    :lines: 62-70
     :language: python
     :caption: Excluding individual routes from ``CustomAuthenticationMiddleware``
 
 And of course use the same kind of mechanism for dependencies:
 
-.. literalinclude:: /examples/security/using_abstract_authentication_middleware/main.py
-    :lines: 53-57
+.. literalinclude:: /examples/security/using_abstract_authentication_middleware.py
+    :lines: 73-77
     :language: python
     :caption: Using ``CustomAuthenticationMiddleware`` in a dependency
